Themes: InCallUI dialpad digits color
11/2016: Port to Nougat's Dialer (from marshmallow's InCallUI) by
KreAch3R

InCallUI's dialpad digits (not the digits' text) color is linked to the
InCallUI theme primary color, on runtime. This is coded as such because
Google wanted to update the InCallUI's theme (primary and primary-dark)
colors in MSIM devices on the fly, based on which SIM card was used to
make the call.
This commit:
* makes this behavior configurable by a boolean
* links the digits color to a color value

This way, the default behavior is preserved. The themer has to set the
"config_dialpadDigitsStaticColor" boolean to true, in order to theme the
color value. If he doesn't change the boolean, the default behavior is
preserved and the digits color works as Google wanted.

Change-Id: I9fc18dc5068f8bab90fbe8b5cfc608f0500a09c8

# Conflicts:
#	res/values/rr_bools.xml